# http/docs/swagger.yaml
basePath: /api
definitions:
...
  messages.createMessageRequest:
    properties:
      content:
        description: Message content
        example: Hello, everyone!
        maxLength: 5000
        minLength: 1
        type: string
      roomId:
        description: Room ID
        example: 550e8400-e29b-41d4-a716-446655440003
        type: string
    type: object
  messages.createMessageResponse:
    properties:
      content:
        description: Message content
        example: Hello, everyone!
        type: string
      id:
        description: Unique message identifier
        example: 550e8400-e29b-41d4-a716-446655440003
        type: string
      roomId:
        description: Room identifier where message was sent
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: Anonymous chat room API with WebSocket support
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Visper API
  version: "1.0"
paths:
...
  /rooms/{roomId}/messages:
    post:
      consumes:
      - application/json
      description: Creates a new message in the specified room and broadcasts it to
        all connected clients. Messages are persisted only if the room is persistent.
      parameters:
      - description: Room ID
        in: path
        name: roomId
        required: true
        type: string
      - description: Message content
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/messages.createMessageRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Message created successfully
          schema:
            $ref: '#/definitions/messages.createMessageResponse'
        "400":
          description: Bad request - validation error or missing room ID
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized - missing authentication or not a member
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Room not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - MemberAuth: []
      summary: Create a new message
      tags:
      - messages
  /rooms/{roomId}/messages/{messageId}:
    delete:
      description: Deletes a message from the room. Only available for persistent
        rooms.
      parameters:
      - description: Room ID
        in: path
        name: roomId
        required: true
        type: string
      - description: Message ID
        in: path
        name: messageId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: Message deleted successfully
        "400":
          description: Bad request - missing room ID or message ID
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized - missing authentication or not a member
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Room or message not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - MemberAuth: []
      summary: Delete a message
      tags:
      - messages
securityDefinitions:
  MemberAuth:
    in: cookie
    name: member_id
    type: apiKey
swagger: "2.0"
...